expanded console message for the confused and curious
Just in case anyone actually uses this but forgets they installed it, then, through e.g. changes to YouTube's HTML, the script stops working or makes a mess; the technically minded enough to have installed the thing might look in the console to see what's wrong, find the message, be reminded of the extension and therefore have an easier time addressing the issue. Aren't I thoughtful? Yes ;)